Extract `ArendSubExprUtils`
